HBASE-27231 FSHLog should retry writing WAL entries when syncs to HDFS failed
