+++ hbase-server/src/main/java/org/apache/hadoop/hbase/regionserver/wal/FSHLog.java
@@ -17,29 +17,16 @@
  */
 package org.apache.hadoop.hbase.regionserver.wal;
 
-import static org.apache.hadoop.hbase.regionserver.wal.WALActionsListener.RollRequestReason.ERROR;
-import static org.apache.hadoop.hbase.regionserver.wal.WALActionsListener.RollRequestReason.LOW_REPLICATION;
-import static org.apache.hadoop.hbase.regionserver.wal.WALActionsListener.RollRequestReason.SIZE;
 import static org.apache.hadoop.hbase.regionserver.wal.WALActionsListener.RollRequestReason.SLOW_SYNC;
 
-import com.lmax.disruptor.BlockingWaitStrategy;
-import com.lmax.disruptor.EventHandler;
-import com.lmax.disruptor.ExceptionHandler;
-import com.lmax.disruptor.LifecycleAware;
-import com.lmax.disruptor.TimeoutException;
-import com.lmax.disruptor.dsl.Disruptor;
-import com.lmax.disruptor.dsl.ProducerType;
-import io.opentelemetry.api.trace.Span;
 import java.io.IOException;
+import java.io.InterruptedIOException;
 import java.io.OutputStream;
 import java.util.Arrays;
 import java.util.List;
 import java.util.concurrent.BlockingQueue;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
+import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.LinkedBlockingQueue;
-import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.FSDataOutputStream;
@@ -47,14 +34,10 @@ import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
 import org.apache.hadoop.hbase.Abortable;
 import org.apache.hadoop.hbase.HConstants;
-import org.apache.hadoop.hbase.client.RegionInfo;
 import org.apache.hadoop.hbase.util.Bytes;
 import org.apache.hadoop.hbase.util.ClassSize;
 import org.apache.hadoop.hbase.util.CommonFSUtils;
-import org.apache.hadoop.hbase.util.Threads;
 import org.apache.hadoop.hbase.wal.FSHLogProvider;
-import org.apache.hadoop.hbase.wal.WALEdit;
-import org.apache.hadoop.hbase.wal.WALKeyImpl;
 import org.apache.hadoop.hbase.wal.WALProvider.Writer;
 import org.apache.hadoop.hdfs.DFSOutputStream;
 import org.apache.hadoop.hdfs.client.HdfsDataOutputStream;
@@ -62,8 +45,6 @@ import org.apache.hadoop.hdfs.protocol.DatanodeInfo;
 import org.apache.yetus.audience.InterfaceAudience;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import org.apache.hbase.thirdparty.com.google.common.util.concurrent.ThreadFactoryBuilder;
 
 /**
  * The original implementation of FSWAL.
@@ -112,8 +93,6 @@ public class FSHLog extends AbstractFSWAL<Writer> {
   private static final String LOW_REPLICATION_ROLL_LIMIT =
     "hbase.regionserver.hlog.lowreplication.rolllimit";
   private static final int DEFAULT_LOW_REPLICATION_ROLL_LIMIT = 5;
-  private static final String ROLL_ERRORS_TOLERATED = "hbase.regionserver.logroll.errors.tolerated";
-  private static final int DEFAULT_ROLL_ERRORS_TOLERATED = 2;
   private static final String SYNCER_COUNT = "hbase.regionserver.hlog.syncer.count";
   private static final int DEFAULT_SYNCER_COUNT = 5;
   private static final String MAX_BATCH_COUNT = "hbase.regionserver.wal.sync.batch.count";
@@ -123,22 +102,10 @@ public class FSHLog extends AbstractFSWAL<Writer> {
     "hbase.wal.fshlog.wait.on.shutdown.seconds";
   private static final int DEFAULT_FSHLOG_WAIT_ON_SHUTDOWN_IN_SECONDS = 5;
 
-  /**
-   * The nexus at which all incoming handlers meet. Does appends and sync with an ordering. Appends
-   * and syncs are each put on the ring which means handlers need to smash up against the ring twice
-   * (can we make it once only? ... maybe not since time to append is so different from time to sync
-   * and sometimes we don't want to sync or we want to async the sync). The ring is where we make
-   * sure of our ordering and it is also where we do batching up of handler sync calls.
-   */
-  private final Disruptor<RingBufferTruck> disruptor;
-
-  /**
-   * This fellow is run by the above appendExecutor service but it is all about batching up appends
-   * and syncs; it may shutdown without cleaning out the last few appends or syncs. To guard against
-   * this, keep a reference to this handler and do explicit close on way out to make sure all
-   * flushed out before we exit.
-   */
-  private final RingBufferEventHandler ringBufferEventHandler;
+  private static final IOException WITER_REPLACED_EXCEPTION =
+    new IOException("Writer was replaced!");
+  private static final IOException WITER_BROKEN_EXCEPTION = new IOException("Wirter was broken!");
+  private static final IOException WAL_CLOSE_EXCEPTION = new IOException("WAL was closed!");
 
   /**
    * FSDataOutputStream associated with the current SequenceFile.writer
@@ -163,39 +130,12 @@ public class FSHLog extends AbstractFSWAL<Writer> {
   // Enable it if the replications recover.
   private volatile boolean lowReplicationRollEnabled = true;
 
-  /** Number of log close errors tolerated before we abort */
-  private final int closeErrorsTolerated;
-
-  private final AtomicInteger closeErrorCount = new AtomicInteger();
-
-  private final int waitOnShutdownInSeconds;
-  private final ExecutorService closeExecutor = Executors.newCachedThreadPool(
-    new ThreadFactoryBuilder().setDaemon(true).setNameFormat("Close-WAL-Writer-%d").build());
-
   /**
-   * Exception handler to pass the disruptor ringbuffer. Same as native implementation only it logs
-   * using our logger instead of java native logger.
+   * Which syncrunner to use next.
    */
-  static class RingBufferExceptionHandler implements ExceptionHandler<RingBufferTruck> {
+  private int syncRunnerIndex = 0;
 
-    @Override
-    public void handleEventException(Throwable ex, long sequence, RingBufferTruck event) {
-      LOG.error("Sequence=" + sequence + ", event=" + event, ex);
-      throw new RuntimeException(ex);
-    }
-
-    @Override
-    public void handleOnStartException(Throwable ex) {
-      LOG.error(ex.toString(), ex);
-      throw new RuntimeException(ex);
-    }
-
-    @Override
-    public void handleOnShutdownException(Throwable ex) {
-      LOG.error(ex.toString(), ex);
-      throw new RuntimeException(ex);
-    }
-  }
+  private final SyncRunner[] syncRunners;
 
   /**
    * Constructor.
@@ -250,29 +190,26 @@ public class FSHLog extends AbstractFSWAL<Writer> {
       conf.getInt(TOLERABLE_LOW_REPLICATION, CommonFSUtils.getDefaultReplication(fs, this.walDir));
     this.lowReplicationRollLimit =
       conf.getInt(LOW_REPLICATION_ROLL_LIMIT, DEFAULT_LOW_REPLICATION_ROLL_LIMIT);
-    this.closeErrorsTolerated = conf.getInt(ROLL_ERRORS_TOLERATED, DEFAULT_ROLL_ERRORS_TOLERATED);
-    this.waitOnShutdownInSeconds =
-      conf.getInt(FSHLOG_WAIT_ON_SHUTDOWN_IN_SECONDS, DEFAULT_FSHLOG_WAIT_ON_SHUTDOWN_IN_SECONDS);
-    // This is the 'writer' -- a single threaded executor. This single thread 'consumes' what is
-    // put on the ring buffer.
-    String hostingThreadName = Thread.currentThread().getName();
-    // Using BlockingWaitStrategy. Stuff that is going on here takes so long it makes no sense
-    // spinning as other strategies do.
-    this.disruptor = new Disruptor<>(RingBufferTruck::new, getPreallocatedEventCount(),
-      new ThreadFactoryBuilder().setNameFormat(hostingThreadName + ".append-pool-%d")
-        .setDaemon(true).setUncaughtExceptionHandler(Threads.LOGGING_EXCEPTION_HANDLER).build(),
-      ProducerType.MULTI, new BlockingWaitStrategy());
+
     // Advance the ring buffer sequence so that it starts from 1 instead of 0,
     // because SyncFuture.NOT_DONE = 0.
-    this.disruptor.getRingBuffer().next();
+
     int syncerCount = conf.getInt(SYNCER_COUNT, DEFAULT_SYNCER_COUNT);
     int maxBatchCount = conf.getInt(MAX_BATCH_COUNT,
       conf.getInt(HConstants.REGION_SERVER_HANDLER_COUNT, DEFAULT_MAX_BATCH_COUNT));
-    this.ringBufferEventHandler = new RingBufferEventHandler(syncerCount, maxBatchCount);
-    this.disruptor.setDefaultExceptionHandler(new RingBufferExceptionHandler());
-    this.disruptor.handleEventsWith(new RingBufferEventHandler[] { this.ringBufferEventHandler });
-    // Starting up threads in constructor is a no no; Interface should have an init call.
-    this.disruptor.start();
+
+    this.createSingleThreadPoolConsumeExecutor("FSHLog", rootDir, prefix);
+
+    this.setWaitOnShutdownInSeconds(
+      conf.getInt(FSHLOG_WAIT_ON_SHUTDOWN_IN_SECONDS, DEFAULT_FSHLOG_WAIT_ON_SHUTDOWN_IN_SECONDS),
+      FSHLOG_WAIT_ON_SHUTDOWN_IN_SECONDS);
+
+    this.syncRunners = new SyncRunner[syncerCount];
+    for (int i = 0; i < syncerCount; i++) {
+      this.syncRunners[i] = new SyncRunner("sync." + i, maxBatchCount);
+      this.syncRunners[i].start();
+    }
+
   }
 
   /**
@@ -315,193 +252,51 @@ public class FSHLog extends AbstractFSWAL<Writer> {
     return writer;
   }
 
-  /**
-   * Used to manufacture race condition reliably. For testing only.
-   * @see #beforeWaitOnSafePoint()
-   */
-  protected void afterCreatingZigZagLatch() {
-  }
-
-  /**
-   * @see #afterCreatingZigZagLatch()
-   */
-  protected void beforeWaitOnSafePoint() {
-  }
-
   @Override
   protected void doAppend(Writer writer, FSWALEntry entry) throws IOException {
     writer.append(entry);
   }
 
   @Override
-  protected void doReplaceWriter(Path oldPath, Path newPath, Writer nextWriter) throws IOException {
-    // Ask the ring buffer writer to pause at a safe point. Once we do this, the writer
-    // thread will eventually pause. An error hereafter needs to release the writer thread
-    // regardless -- hence the finally block below. Note, this method is called from the FSHLog
-    // constructor BEFORE the ring buffer is set running so it is null on first time through
-    // here; allow for that.
-    SyncFuture syncFuture = null;
-    SafePointZigZagLatch zigzagLatch = null;
-    long sequence = -1L;
-    if (this.writer != null && this.ringBufferEventHandler != null) {
-      // Get sequence first to avoid dead lock when ring buffer is full
-      // Considering below sequence
-      // 1. replaceWriter is called and zigzagLatch is initialized
-      // 2. ringBufferEventHandler#onEvent is called and arrives at #attainSafePoint(long) then wait
-      // on safePointReleasedLatch
-      // 3. Since ring buffer is full, if we get sequence when publish sync, the replaceWriter
-      // thread will wait for the ring buffer to be consumed, but the only consumer is waiting
-      // replaceWriter thread to release safePointReleasedLatch, which causes a deadlock
-      sequence = getSequenceOnRingBuffer();
-      zigzagLatch = this.ringBufferEventHandler.attainSafePoint();
-    }
-    afterCreatingZigZagLatch();
-    try {
-      // Wait on the safe point to be achieved. Send in a sync in case nothing has hit the
-      // ring buffer between the above notification of writer that we want it to go to
-      // 'safe point' and then here where we are waiting on it to attain safe point. Use
-      // 'sendSync' instead of 'sync' because we do not want this thread to block waiting on it
-      // to come back. Cleanup this syncFuture down below after we are ready to run again.
-      try {
-        if (zigzagLatch != null) {
-          // use assert to make sure no change breaks the logic that
-          // sequence and zigzagLatch will be set together
-          assert sequence > 0L : "Failed to get sequence from ring buffer";
-          syncFuture = zigzagLatch.waitSafePoint(publishSyncOnRingBuffer(sequence, false));
-        }
-      } catch (FailedSyncBeforeLogCloseException e) {
-        // If unflushed/unsynced entries on close, it is reason to abort.
-        if (isUnflushedEntries()) {
-          throw e;
-        }
-        LOG.warn(
-          "Failed sync-before-close but no outstanding appends; closing WAL" + e.getMessage());
-      }
-      long oldFileLen = 0L;
-      // It is at the safe point. Swap out writer from under the blocked writer thread.
-      if (this.writer != null) {
-        oldFileLen = this.writer.getLength();
-        // In case of having unflushed entries or we already reached the
-        // closeErrorsTolerated count, call the closeWriter inline rather than in async
-        // way so that in case of an IOE we will throw it back and abort RS.
-        inflightWALClosures.put(oldPath.getName(), writer);
-        if (isUnflushedEntries() || closeErrorCount.get() >= this.closeErrorsTolerated) {
-          closeWriter(this.writer, oldPath, true);
-        } else {
-          Writer localWriter = this.writer;
-          closeExecutor.execute(() -> {
-            try {
-              closeWriter(localWriter, oldPath, false);
-            } catch (IOException e) {
-              // We will never reach here.
-            }
-          });
-        }
-      }
-      logRollAndSetupWalProps(oldPath, newPath, oldFileLen);
-      this.writer = nextWriter;
-      if (nextWriter != null && nextWriter instanceof ProtobufLogWriter) {
-        this.hdfs_out = ((ProtobufLogWriter) nextWriter).getStream();
-      } else {
-        this.hdfs_out = null;
-      }
-    } catch (InterruptedException ie) {
-      // Perpetuate the interrupt
-      Thread.currentThread().interrupt();
-    } catch (IOException e) {
-      long count = getUnflushedEntriesCount();
-      LOG.error("Failed close of WAL writer " + oldPath + ", unflushedEntries=" + count, e);
-      throw new FailedLogCloseException(oldPath + ", unflushedEntries=" + count, e);
-    } finally {
-      // Let the writer thread go regardless, whether error or not.
-      if (zigzagLatch != null) {
-        // Reset rollRequested status
-        rollRequested.set(false);
-        zigzagLatch.releaseSafePoint();
-        // syncFuture will be null if we failed our wait on safe point above. Otherwise, if
-        // latch was obtained successfully, the sync we threw in either trigger the latch or it
-        // got stamped with an exception because the WAL was damaged and we could not sync. Now
-        // the write pipeline has been opened up again by releasing the safe point, process the
-        // syncFuture we got above. This is probably a noop but it may be stale exception from
-        // when old WAL was in place. Catch it if so.
-        if (syncFuture != null) {
-          try {
-            blockOnSync(syncFuture);
-          } catch (IOException ioe) {
-            if (LOG.isTraceEnabled()) {
-              LOG.trace("Stale sync exception", ioe);
-            }
-          }
-        }
-      }
-    }
-  }
-
-  private void closeWriter(Writer writer, Path path, boolean syncCloseCall) throws IOException {
-    Span span = Span.current();
-    try {
-      span.addEvent("closing writer");
-      writer.close();
-      span.addEvent("writer closed");
-    } catch (IOException ioe) {
-      int errors = closeErrorCount.incrementAndGet();
-      boolean hasUnflushedEntries = isUnflushedEntries();
-      if (syncCloseCall && (hasUnflushedEntries || (errors > this.closeErrorsTolerated))) {
-        LOG.error("Close of WAL " + path + " failed. Cause=\"" + ioe.getMessage() + "\", errors="
-          + errors + ", hasUnflushedEntries=" + hasUnflushedEntries);
-        throw ioe;
-      }
-      LOG.warn("Riding over failed WAL close of " + path
-        + "; THIS FILE WAS NOT CLOSED BUT ALL EDITS SYNCED SO SHOULD BE OK", ioe);
-    } finally {
-      inflightWALClosures.remove(path.getName());
+  protected void onWriterReplaced(Writer nextWriter) {
+    if (nextWriter != null && nextWriter instanceof ProtobufLogWriter) {
+      this.hdfs_out = ((ProtobufLogWriter) nextWriter).getStream();
+    } else {
+      this.hdfs_out = null;
     }
   }
 
   @Override
-  protected void doShutdown() throws IOException {
-    // Shutdown the disruptor. Will stop after all entries have been processed. Make sure we
-    // have stopped incoming appends before calling this else it will not shutdown. We are
-    // conservative below waiting a long time and if not elapsed, then halting.
-    if (this.disruptor != null) {
-      long timeoutms = conf.getLong("hbase.wal.disruptor.shutdown.timeout.ms", 60000);
-      try {
-        this.disruptor.shutdown(timeoutms, TimeUnit.MILLISECONDS);
-      } catch (TimeoutException e) {
-        LOG.warn("Timed out bringing down disruptor after " + timeoutms + "ms; forcing halt "
-          + "(It is a problem if this is NOT an ABORT! -- DATALOSS!!!!)");
-        this.disruptor.halt();
-        this.disruptor.shutdown();
-      }
+  protected void doCleanUp() {
+    for (SyncRunner syncRunner : this.syncRunners) {
+      syncRunner.shutDown();
     }
-
-    if (LOG.isDebugEnabled()) {
-      LOG.debug("Closing WAL writer in " + CommonFSUtils.getPath(walDir));
-    }
-    if (this.writer != null) {
-      this.writer.close();
-      this.writer = null;
-    }
-    closeExecutor.shutdown();
-    try {
-      if (!closeExecutor.awaitTermination(waitOnShutdownInSeconds, TimeUnit.SECONDS)) {
-        LOG.error(
-          "We have waited {} seconds but the close of writer(s) doesn't complete."
-            + "Please check the status of underlying filesystem"
-            + " or increase the wait time by the config \"{}\"",
-          this.waitOnShutdownInSeconds, FSHLOG_WAIT_ON_SHUTDOWN_IN_SECONDS);
-      }
-    } catch (InterruptedException e) {
-      LOG.error("The wait for termination of FSHLog writer(s) is interrupted");
-      Thread.currentThread().interrupt();
-    }
-  }
+  };
 
   @Override
-  protected long append(final RegionInfo hri, final WALKeyImpl key, final WALEdit edits,
-    final boolean inMemstore) throws IOException {
-    return stampSequenceIdAndPublishToRingBuffer(hri, key, edits, inMemstore,
-      disruptor.getRingBuffer());
+  protected CompletableFuture<Long> doWriterSync(Writer writer, boolean shouldUseHSync,
+    long txidWhenSync) {
+    CompletableFuture<Long> future = new CompletableFuture<>();
+    SyncRequest syncRequest = new SyncRequest(writer, shouldUseHSync, txidWhenSync, future);
+    this.syncRunnerIndex = (this.syncRunnerIndex + 1) % this.syncRunners.length;
+    this.syncRunners[this.syncRunnerIndex].offer(syncRequest);
+    return future;
+  }
+
+  class SyncRequest {
+    private final Writer writer;
+    private final boolean shouldUseHSync;
+    private final long sequenceWhenSync;
+    private final CompletableFuture<Long> completableFuture;
+
+    public SyncRequest(Writer writer, boolean shouldUseHSync, long txidWhenSync,
+      CompletableFuture<Long> completableFuture) {
+      this.writer = writer;
+      this.shouldUseHSync = shouldUseHSync;
+      this.sequenceWhenSync = txidWhenSync;
+      this.completableFuture = completableFuture;
+    }
+
   }
 
   /**
@@ -521,10 +316,9 @@ public class FSHLog extends AbstractFSWAL<Writer> {
    * completes.
    */
   private class SyncRunner extends Thread {
-    private volatile long sequence;
     // Keep around last exception thrown. Clear on successful sync.
-    private final BlockingQueue<SyncFuture> syncFutures;
-    private volatile SyncFuture takeSyncFuture = null;
+    private final BlockingQueue<SyncRequest> syncRequests;
+    private volatile boolean shutDown = false;
 
     SyncRunner(final String name, final int maxHandlersCount) {
       super(name);
@@ -541,183 +335,158 @@ public class FSHLog extends AbstractFSWAL<Writer> {
       // the meta table when succesful (i.e. sync), closing handlers -- etc. These are usually
       // much fewer in number than the user-space handlers so Q-size should be user handlers plus
       // some space for these other handlers. Lets multiply by 3 for good-measure.
-      this.syncFutures = new LinkedBlockingQueue<>(maxHandlersCount * 3);
+      this.syncRequests = new LinkedBlockingQueue<>(maxHandlersCount * 3);
     }
 
-    void offer(final long sequence, final SyncFuture[] syncFutures, final int syncFutureCount) {
-      // Set sequence first because the add to the queue will wake the thread if sleeping.
-      this.sequence = sequence;
-      for (int i = 0; i < syncFutureCount; ++i) {
-        this.syncFutures.add(syncFutures[i]);
+    void offer(SyncRequest syncRequest) {
+      if (this.shutDown) {
+        syncRequest.completableFuture.completeExceptionally(WAL_CLOSE_EXCEPTION);
+        return;
+      }
+      try {
+        this.syncRequests.put(syncRequest);
+      } catch (InterruptedException e) {
+        LOG.warn("interrupted when put syncRequest", e);
+        syncRequest.completableFuture
+          .completeExceptionally(new InterruptedIOException("interrupted when put syncRequest"));
+        return;
+      }
+
+      // recheck
+      if (this.shutDown) {
+        if (this.syncRequests.remove(syncRequest)) {
+          syncRequest.completableFuture.completeExceptionally(WAL_CLOSE_EXCEPTION);
+        }
       }
     }
 
-    /**
-     * Release the passed <code>syncFuture</code>
-     * @return Returns 1.
-     */
-    private int releaseSyncFuture(final SyncFuture syncFuture, final long currentSequence,
-      final Throwable t) {
-      if (!syncFuture.done(currentSequence, t)) {
-        throw new IllegalStateException();
+    private void completeSyncRequests(SyncRequest syncRequest, long syncedSequenceId) {
+      if (syncRequest != null) {
+        syncRequest.completableFuture.complete(syncedSequenceId);
       }
-
-      // This function releases one sync future only.
-      return 1;
-    }
-
-    /**
-     * Release all SyncFutures whose sequence is <= <code>currentSequence</code>.
-     * @param t May be non-null if we are processing SyncFutures because an exception was thrown.
-     * @return Count of SyncFutures we let go.
-     */
-    private int releaseSyncFutures(final long currentSequence, final Throwable t) {
-      int syncCount = 0;
-      for (SyncFuture syncFuture; (syncFuture = this.syncFutures.peek()) != null;) {
-        if (syncFuture.getTxid() > currentSequence) {
+      while (true) {
+        SyncRequest head = this.syncRequests.peek();
+        if (head == null) {
           break;
         }
-        releaseSyncFuture(syncFuture, currentSequence, t);
-        if (!this.syncFutures.remove(syncFuture)) {
-          throw new IllegalStateException(syncFuture.toString());
-        }
-        syncCount++;
-      }
-      return syncCount;
-    }
-
-    /**
-     * @param sequence The sequence we ran the filesystem sync against.
-     * @return Current highest synced sequence.
-     */
-    private long updateHighestSyncedSequence(long sequence) {
-      long currentHighestSyncedSequence;
-      // Set the highestSyncedSequence IFF our current sequence id is the 'highest'.
-      do {
-        currentHighestSyncedSequence = highestSyncedTxid.get();
-        if (currentHighestSyncedSequence >= sequence) {
-          // Set the sync number to current highwater mark; might be able to let go more
-          // queued sync futures
-          sequence = currentHighestSyncedSequence;
+        if (head.sequenceWhenSync > syncedSequenceId) {
           break;
         }
-      } while (!highestSyncedTxid.compareAndSet(currentHighestSyncedSequence, sequence));
-      return sequence;
+        head.completableFuture.complete(syncedSequenceId);
+        this.syncRequests.poll();
+      }
     }
 
-    boolean areSyncFuturesReleased() {
-      // check whether there is no sync futures offered, and no in-flight sync futures that is being
-      // processed.
-      return syncFutures.size() <= 0 && takeSyncFuture == null;
+    private void completeExceptionallySyncRequests(SyncRequest syncRequest, Exception exception) {
+      if (syncRequest != null) {
+        syncRequest.completableFuture.completeExceptionally(exception);
+      }
+      while (true) {
+        SyncRequest head = this.syncRequests.peek();
+        if (head == null) {
+          break;
+        }
+        if (head.writer != syncRequest.writer) {
+          break;
+        }
+        head.completableFuture.completeExceptionally(exception);
+        this.syncRequests.poll();
+      }
+    }
+
+    private SyncRequest takeSyncRequest() throws InterruptedException {
+      while (true) {
+        // We have to process what we 'take' from the queue
+        SyncRequest syncRequest = this.syncRequests.take();
+        // See if we can process any syncfutures BEFORE we go sync.
+        long currentHighestSyncedSequence = highestSyncedTxid.get();
+        if (syncRequest.sequenceWhenSync < currentHighestSyncedSequence) {
+          syncRequest.completableFuture.complete(currentHighestSyncedSequence);
+          continue;
+        }
+        return syncRequest;
+      }
     }
 
     @Override
     public void run() {
-      long currentSequence;
-      while (!isInterrupted()) {
-        int syncCount = 0;
-
+      while (!this.shutDown) {
         try {
-          // Make a local copy of takeSyncFuture after we get it. We've been running into NPEs
-          // 2020-03-22 16:54:32,180 WARN [sync.1] wal.FSHLog$SyncRunner(589): UNEXPECTED
-          // java.lang.NullPointerException
-          // at org.apache.hadoop.hbase.regionserver.wal.FSHLog$SyncRunner.run(FSHLog.java:582)
-          // at java.lang.Thread.run(Thread.java:748)
-          SyncFuture sf;
-          while (true) {
-            takeSyncFuture = null;
-            // We have to process what we 'take' from the queue
-            takeSyncFuture = this.syncFutures.take();
-            // Make local copy.
-            sf = takeSyncFuture;
-            currentSequence = this.sequence;
-            long syncFutureSequence = sf.getTxid();
-            if (syncFutureSequence > currentSequence) {
-              throw new IllegalStateException("currentSequence=" + currentSequence
-                + ", syncFutureSequence=" + syncFutureSequence);
-            }
-            // See if we can process any syncfutures BEFORE we go sync.
-            long currentHighestSyncedSequence = highestSyncedTxid.get();
-            if (currentSequence < currentHighestSyncedSequence) {
-              syncCount += releaseSyncFuture(sf, currentHighestSyncedSequence, null);
-              // Done with the 'take'. Go around again and do a new 'take'.
-              continue;
-            }
-            break;
-          }
+          SyncRequest syncRequest = this.takeSyncRequest();
           // I got something. Lets run. Save off current sequence number in case it changes
           // while we run.
-          long start = System.nanoTime();
-          Throwable lastException = null;
+          long currentSequenceToUse = syncRequest.sequenceWhenSync;
+          boolean writerBroken = isWriterBroken();
+          long currentHighestProcessedAppendTxid = highestProcessedAppendTxid;
+          Writer currentWriter = writer;
+          if (currentWriter != syncRequest.writer) {
+            syncRequest.completableFuture.completeExceptionally(WITER_REPLACED_EXCEPTION);
+            continue;
+          }
+          if (writerBroken) {
+            syncRequest.completableFuture.completeExceptionally(WITER_BROKEN_EXCEPTION);
+            continue;
+          }
+          if (currentHighestProcessedAppendTxid > currentSequenceToUse) {
+            currentSequenceToUse = currentHighestProcessedAppendTxid;
+          }
+          Exception lastException = null;
           try {
-            long unSyncedFlushSeq = highestUnsyncedTxid;
-            writer.sync(sf.isForceSync());
-            if (unSyncedFlushSeq > currentSequence) {
-              currentSequence = unSyncedFlushSeq;
-            }
-            currentSequence = updateHighestSyncedSequence(currentSequence);
+            writer.sync(syncRequest.shouldUseHSync);
           } catch (IOException e) {
-            LOG.error("Error syncing, request close of WAL", e);
+            LOG.error("Error syncing", e);
             lastException = e;
           } catch (Exception e) {
             LOG.warn("UNEXPECTED", e);
             lastException = e;
           } finally {
-            // First release what we 'took' from the queue.
-            syncCount += releaseSyncFuture(takeSyncFuture, currentSequence, lastException);
-            // Can we release other syncs?
-            syncCount += releaseSyncFutures(currentSequence, lastException);
             if (lastException != null) {
-              requestLogRoll(ERROR);
+              this.completeExceptionallySyncRequests(syncRequest, lastException);
             } else {
-              checkLogRoll();
+              this.completeSyncRequests(syncRequest, currentSequenceToUse);
             }
           }
-          postSync(System.nanoTime() - start, syncCount);
         } catch (InterruptedException e) {
           // Presume legit interrupt.
-          Thread.currentThread().interrupt();
+          LOG.info("interrupted");
         } catch (Throwable t) {
           LOG.warn("UNEXPECTED, continuing", t);
         }
       }
+      this.clearSyncRequestsWhenShutDown();
+    }
+
+    private void clearSyncRequestsWhenShutDown() {
+      while (true) {
+        SyncRequest syncRequest = this.syncRequests.poll();
+        if (syncRequest == null) {
+          break;
+        }
+        syncRequest.completableFuture.completeExceptionally(WAL_CLOSE_EXCEPTION);
+      }
+    }
+
+    void shutDown() {
+      try {
+        this.shutDown = true;
+        this.interrupt();
+        this.join();
+      } catch (InterruptedException e) {
+        LOG.warn("interrupted", e);
+        Thread.currentThread().interrupt();
+      }
     }
   }
 
-  /**
-   * Schedule a log roll if needed.
-   */
-  private boolean checkLogRoll() {
-    // If we have already requested a roll, do nothing
+  @Override
+  protected void checkSlowSyncCount() {
     if (isLogRollRequested()) {
-      return false;
+      return;
     }
-    // Will return immediately if we are in the middle of a WAL log roll currently.
-    if (!rollWriterLock.tryLock()) {
-      return false;
+    if (doCheckSlowSync()) {
+      // We log this already in checkSlowSync
+      requestLogRoll(SLOW_SYNC);
     }
-    try {
-      if (doCheckLogLowReplication()) {
-        LOG.warn("Requesting log roll because of low replication, current pipeline: "
-          + Arrays.toString(getPipeline()));
-        requestLogRoll(LOW_REPLICATION);
-        return true;
-      } else if (writer != null && writer.getLength() > logrollsize) {
-        if (LOG.isDebugEnabled()) {
-          LOG.debug("Requesting log roll because of file size threshold; length="
-            + writer.getLength() + ", logrollsize=" + logrollsize);
-        }
-        requestLogRoll(SIZE);
-        return true;
-      } else if (doCheckSlowSync()) {
-        // We log this already in checkSlowSync
-        requestLogRoll(SLOW_SYNC);
-        return true;
-      }
-    } finally {
-      rollWriterLock.unlock();
-    }
-    return false;
   }
 
   /** Returns true if number of replicas for the WAL is lower than threshold */
@@ -767,33 +536,6 @@ public class FSHLog extends AbstractFSWAL<Writer> {
     return logRollNeeded;
   }
 
-  protected long getSequenceOnRingBuffer() {
-    return this.disruptor.getRingBuffer().next();
-  }
-
-  private SyncFuture publishSyncOnRingBuffer(boolean forceSync) {
-    long sequence = getSequenceOnRingBuffer();
-    return publishSyncOnRingBuffer(sequence, forceSync);
-  }
-
-  protected SyncFuture publishSyncOnRingBuffer(long sequence, boolean forceSync) {
-    // here we use ring buffer sequence as transaction id
-    SyncFuture syncFuture = getSyncFuture(sequence, forceSync);
-    try {
-      RingBufferTruck truck = this.disruptor.getRingBuffer().get(sequence);
-      truck.load(syncFuture);
-    } finally {
-      this.disruptor.getRingBuffer().publish(sequence);
-    }
-    return syncFuture;
-  }
-
-  // Sync all known transactions
-  private void publishSyncThenBlockOnCompletion(boolean forceSync) throws IOException {
-    SyncFuture syncFuture = publishSyncOnRingBuffer(forceSync);
-    blockOnSync(syncFuture);
-  }
-
   /**
    * {@inheritDoc}
    * <p>
@@ -814,20 +556,6 @@ public class FSHLog extends AbstractFSWAL<Writer> {
     return 0;
   }
 
-  @Override
-  protected void doSync(boolean forceSync) throws IOException {
-    publishSyncThenBlockOnCompletion(forceSync);
-  }
-
-  @Override
-  protected void doSync(long txid, boolean forceSync) throws IOException {
-    if (this.highestSyncedTxid.get() >= txid) {
-      // Already sync'd.
-      return;
-    }
-    publishSyncThenBlockOnCompletion(forceSync);
-  }
-
   boolean isLowReplicationRollEnabled() {
     return lowReplicationRollEnabled;
   }
@@ -835,361 +563,6 @@ public class FSHLog extends AbstractFSWAL<Writer> {
   public static final long FIXED_OVERHEAD =
     ClassSize.align(ClassSize.OBJECT + (5 * ClassSize.REFERENCE) + (2 * ClassSize.ATOMIC_INTEGER)
       + (3 * Bytes.SIZEOF_INT) + (4 * Bytes.SIZEOF_LONG));
-
-  /**
-   * This class is used coordinating two threads holding one thread at a 'safe point' while the
-   * orchestrating thread does some work that requires the first thread paused: e.g. holding the WAL
-   * writer while its WAL is swapped out from under it by another thread.
-   * <p>
-   * Thread A signals Thread B to hold when it gets to a 'safe point'. Thread A wait until Thread B
-   * gets there. When the 'safe point' has been attained, Thread B signals Thread A. Thread B then
-   * holds at the 'safe point'. Thread A on notification that Thread B is paused, goes ahead and
-   * does the work it needs to do while Thread B is holding. When Thread A is done, it flags B and
-   * then Thread A and Thread B continue along on their merry way. Pause and signalling 'zigzags'
-   * between the two participating threads. We use two latches -- one the inverse of the other --
-   * pausing and signaling when states are achieved.
-   * <p>
-   * To start up the drama, Thread A creates an instance of this class each time it would do this
-   * zigzag dance and passes it to Thread B (these classes use Latches so it is one shot only).
-   * Thread B notices the new instance (via reading a volatile reference or how ever) and it starts
-   * to work toward the 'safe point'. Thread A calls {@link #waitSafePoint(SyncFuture)} when it
-   * cannot proceed until the Thread B 'safe point' is attained. Thread A will be held inside in
-   * {@link #waitSafePoint(SyncFuture)} until Thread B reaches the 'safe point'. Once there, Thread
-   * B frees Thread A by calling {@link #safePointAttained()}. Thread A now knows Thread B is at the
-   * 'safe point' and that it is holding there (When Thread B calls {@link #safePointAttained()} it
-   * blocks here until Thread A calls {@link #releaseSafePoint()}). Thread A proceeds to do what it
-   * needs to do while Thread B is paused. When finished, it lets Thread B lose by calling
-   * {@link #releaseSafePoint()} and away go both Threads again.
-   */
-  static class SafePointZigZagLatch {
-    /**
-     * Count down this latch when safe point attained.
-     */
-    private volatile CountDownLatch safePointAttainedLatch = new CountDownLatch(1);
-    /**
-     * Latch to wait on. Will be released when we can proceed.
-     */
-    private volatile CountDownLatch safePointReleasedLatch = new CountDownLatch(1);
-
-    private void checkIfSyncFailed(SyncFuture syncFuture) throws FailedSyncBeforeLogCloseException {
-      Throwable t = syncFuture.getThrowable();
-      if (t != null) {
-        throw new FailedSyncBeforeLogCloseException(t);
-      }
-    }
-
-    /**
-     * For Thread A to call when it is ready to wait on the 'safe point' to be attained. Thread A
-     * will be held in here until Thread B calls {@link #safePointAttained()}
-     * @param syncFuture We need this as barometer on outstanding syncs. If it comes home with an
-     *                   exception, then something is up w/ our syncing.
-     * @return The passed <code>syncFuture</code>
-     */
-    SyncFuture waitSafePoint(SyncFuture syncFuture)
-      throws InterruptedException, FailedSyncBeforeLogCloseException {
-      while (!this.safePointAttainedLatch.await(1, TimeUnit.MILLISECONDS)) {
-        checkIfSyncFailed(syncFuture);
-      }
-      checkIfSyncFailed(syncFuture);
-      return syncFuture;
-    }
-
-    /** Returns if the safepoint has been attained. */
-    @InterfaceAudience.Private
-    boolean isSafePointAttained() {
-      return this.safePointAttainedLatch.getCount() == 0;
-    }
-
-    /**
-     * Called by Thread B when it attains the 'safe point'. In this method, Thread B signals Thread
-     * A it can proceed. Thread B will be held in here until {@link #releaseSafePoint()} is called
-     * by Thread A.
-     */
-    void safePointAttained() throws InterruptedException {
-      this.safePointAttainedLatch.countDown();
-      this.safePointReleasedLatch.await();
-    }
-
-    /**
-     * Called by Thread A when it is done with the work it needs to do while Thread B is halted.
-     * This will release the Thread B held in a call to {@link #safePointAttained()}
-     */
-    void releaseSafePoint() {
-      this.safePointReleasedLatch.countDown();
-    }
-
-    /** Returns True is this is a 'cocked', fresh instance, and not one that has already fired. */
-    boolean isCocked() {
-      return this.safePointAttainedLatch.getCount() > 0
-        && this.safePointReleasedLatch.getCount() > 0;
-    }
-  }
-
-  /**
-   * Handler that is run by the disruptor ringbuffer consumer. Consumer is a SINGLE
-   * 'writer/appender' thread. Appends edits and starts up sync runs. Tries its best to batch up
-   * syncs. There is no discernible benefit batching appends so we just append as they come in
-   * because it simplifies the below implementation. See metrics for batching effectiveness (In
-   * measurement, at 100 concurrent handlers writing 1k, we are batching > 10 appends and 10 handler
-   * sync invocations for every actual dfsclient sync call; at 10 concurrent handlers, YMMV).
-   * <p>
-   * Herein, we have an array into which we store the sync futures as they come in. When we have a
-   * 'batch', we'll then pass what we have collected to a SyncRunner thread to do the filesystem
-   * sync. When it completes, it will then call {@link SyncFuture#done(long, Throwable)} on each of
-   * SyncFutures in the batch to release blocked Handler threads.
-   * <p>
-   * I've tried various effects to try and make latencies low while keeping throughput high. I've
-   * tried keeping a single Queue of SyncFutures in this class appending to its tail as the syncs
-   * coming and having sync runner threads poll off the head to 'finish' completed SyncFutures. I've
-   * tried linkedlist, and various from concurrent utils whether LinkedBlockingQueue or
-   * ArrayBlockingQueue, etc. The more points of synchronization, the more 'work' (according to
-   * 'perf stats') that has to be done; small increases in stall percentages seem to have a big
-   * impact on throughput/latencies. The below model where we have an array into which we stash the
-   * syncs and then hand them off to the sync thread seemed like a decent compromise. See HBASE-8755
-   * for more detail.
-   */
-  class RingBufferEventHandler implements EventHandler<RingBufferTruck>, LifecycleAware {
-    private final SyncRunner[] syncRunners;
-    private final SyncFuture[] syncFutures;
-    // Had 'interesting' issues when this was non-volatile. On occasion, we'd not pass all
-    // syncFutures to the next sync'ing thread.
-    private AtomicInteger syncFuturesCount = new AtomicInteger();
-    private volatile SafePointZigZagLatch zigzagLatch;
-    /**
-     * Set if we get an exception appending or syncing so that all subsequence appends and syncs on
-     * this WAL fail until WAL is replaced.
-     */
-    private Exception exception = null;
-    /**
-     * Object to block on while waiting on safe point.
-     */
-    private final Object safePointWaiter = new Object();
-    private volatile boolean shutdown = false;
-
-    /**
-     * Which syncrunner to use next.
-     */
-    private int syncRunnerIndex;
-
-    RingBufferEventHandler(final int syncRunnerCount, final int maxBatchCount) {
-      this.syncFutures = new SyncFuture[maxBatchCount];
-      this.syncRunners = new SyncRunner[syncRunnerCount];
-      for (int i = 0; i < syncRunnerCount; i++) {
-        this.syncRunners[i] = new SyncRunner("sync." + i, maxBatchCount);
-      }
-    }
-
-    private void cleanupOutstandingSyncsOnException(final long sequence, final Exception e) {
-      // There could be handler-count syncFutures outstanding.
-      for (int i = 0; i < this.syncFuturesCount.get(); i++) {
-        this.syncFutures[i].done(sequence, e);
-      }
-      offerDoneSyncsBackToCache();
-    }
-
-    /**
-     * Offers the finished syncs back to the cache for reuse.
-     */
-    private void offerDoneSyncsBackToCache() {
-      for (int i = 0; i < this.syncFuturesCount.get(); i++) {
-        syncFutureCache.offer(syncFutures[i]);
-      }
-      this.syncFuturesCount.set(0);
-    }
-
-    /** Returns True if outstanding sync futures still */
-    private boolean isOutstandingSyncs() {
-      // Look at SyncFutures in the EventHandler
-      for (int i = 0; i < this.syncFuturesCount.get(); i++) {
-        if (!this.syncFutures[i].isDone()) {
-          return true;
-        }
-      }
-
-      return false;
-    }
-
-    private boolean isOutstandingSyncsFromRunners() {
-      // Look at SyncFutures in the SyncRunners
-      for (SyncRunner syncRunner : syncRunners) {
-        if (syncRunner.isAlive() && !syncRunner.areSyncFuturesReleased()) {
-          return true;
-        }
-      }
-      return false;
-    }
-
-    @Override
-    // We can set endOfBatch in the below method if at end of our this.syncFutures array
-    public void onEvent(final RingBufferTruck truck, final long sequence, boolean endOfBatch)
-      throws Exception {
-      // Appends and syncs are coming in order off the ringbuffer. We depend on this fact. We'll
-      // add appends to dfsclient as they come in. Batching appends doesn't give any significant
-      // benefit on measurement. Handler sync calls we will batch up. If we get an exception
-      // appending an edit, we fail all subsequent appends and syncs with the same exception until
-      // the WAL is reset. It is important that we not short-circuit and exit early this method.
-      // It is important that we always go through the attainSafePoint on the end. Another thread,
-      // the log roller may be waiting on a signal from us here and will just hang without it.
-
-      try {
-        if (truck.type() == RingBufferTruck.Type.SYNC) {
-          this.syncFutures[this.syncFuturesCount.getAndIncrement()] = truck.unloadSync();
-          // Force flush of syncs if we are carrying a full complement of syncFutures.
-          if (this.syncFuturesCount.get() == this.syncFutures.length) {
-            endOfBatch = true;
-          }
-        } else if (truck.type() == RingBufferTruck.Type.APPEND) {
-          FSWALEntry entry = truck.unloadAppend();
-          try {
-            if (this.exception != null) {
-              // Return to keep processing events coming off the ringbuffer
-              return;
-            }
-            append(entry);
-          } catch (Exception e) {
-            // Failed append. Record the exception.
-            this.exception = e;
-            // invoking cleanupOutstandingSyncsOnException when append failed with exception,
-            // it will cleanup existing sync requests recorded in syncFutures but not offered to
-            // SyncRunner yet,
-            // so there won't be any sync future left over if no further truck published to
-            // disruptor.
-            cleanupOutstandingSyncsOnException(sequence,
-              this.exception instanceof DamagedWALException
-                ? this.exception
-                : new DamagedWALException("On sync", this.exception));
-            // Return to keep processing events coming off the ringbuffer
-            return;
-          } finally {
-            entry.release();
-          }
-        } else {
-          // What is this if not an append or sync. Fail all up to this!!!
-          cleanupOutstandingSyncsOnException(sequence,
-            new IllegalStateException("Neither append nor sync"));
-          // Return to keep processing.
-          return;
-        }
-
-        // TODO: Check size and if big go ahead and call a sync if we have enough data.
-        // This is a sync. If existing exception, fall through. Else look to see if batch.
-        if (this.exception == null) {
-          // If not a batch, return to consume more events from the ring buffer before proceeding;
-          // we want to get up a batch of syncs and appends before we go do a filesystem sync.
-          if (!endOfBatch || this.syncFuturesCount.get() <= 0) {
-            return;
-          }
-          // syncRunnerIndex is bound to the range [0, Integer.MAX_INT - 1] as follows:
-          // * The maximum value possible for syncRunners.length is Integer.MAX_INT
-          // * syncRunnerIndex starts at 0 and is incremented only here
-          // * after the increment, the value is bounded by the '%' operator to
-          // [0, syncRunners.length), presuming the value was positive prior to
-          // the '%' operator.
-          // * after being bound to [0, Integer.MAX_INT - 1], the new value is stored in
-          // syncRunnerIndex ensuring that it can't grow without bound and overflow.
-          // * note that the value after the increment must be positive, because the most it
-          // could have been prior was Integer.MAX_INT - 1 and we only increment by 1.
-          this.syncRunnerIndex = (this.syncRunnerIndex + 1) % this.syncRunners.length;
-          try {
-            // Below expects that the offer 'transfers' responsibility for the outstanding syncs to
-            // the syncRunner. We should never get an exception in here.
-            this.syncRunners[this.syncRunnerIndex].offer(sequence, this.syncFutures,
-              this.syncFuturesCount.get());
-          } catch (Exception e) {
-            // Should NEVER get here.
-            requestLogRoll(ERROR);
-            this.exception = new DamagedWALException("Failed offering sync", e);
-          }
-        }
-        // We may have picked up an exception above trying to offer sync
-        if (this.exception != null) {
-          cleanupOutstandingSyncsOnException(sequence,
-            this.exception instanceof DamagedWALException
-              ? this.exception
-              : new DamagedWALException("On sync", this.exception));
-        }
-        attainSafePoint(sequence);
-        // It is critical that we offer the futures back to the cache for reuse here after the
-        // safe point is attained and all the clean up has been done. There have been
-        // issues with reusing sync futures early causing WAL lockups, see HBASE-25984.
-        offerDoneSyncsBackToCache();
-      } catch (Throwable t) {
-        LOG.error("UNEXPECTED!!! syncFutures.length=" + this.syncFutures.length, t);
-      }
-    }
-
-    SafePointZigZagLatch attainSafePoint() {
-      this.zigzagLatch = new SafePointZigZagLatch();
-      return this.zigzagLatch;
-    }
-
-    /**
-     * Check if we should attain safe point. If so, go there and then wait till signalled before we
-     * proceeding.
-     */
-    private void attainSafePoint(final long currentSequence) {
-      if (this.zigzagLatch == null || !this.zigzagLatch.isCocked()) {
-        return;
-      }
-      // If here, another thread is waiting on us to get to safe point. Don't leave it hanging.
-      beforeWaitOnSafePoint();
-      try {
-        // Wait on outstanding syncers; wait for them to finish syncing (unless we've been
-        // shutdown or unless our latch has been thrown because we have been aborted or unless
-        // this WAL is broken and we can't get a sync/append to complete).
-        while (
-          (!this.shutdown && this.zigzagLatch.isCocked()
-            && highestSyncedTxid.get() < currentSequence &&
-            // We could be in here and all syncs are failing or failed. Check for this. Otherwise
-            // we'll just be stuck here for ever. In other words, ensure there syncs running.
-            isOutstandingSyncs())
-            // Wait for all SyncRunners to finish their work so that we can replace the writer
-            || isOutstandingSyncsFromRunners()
-        ) {
-          synchronized (this.safePointWaiter) {
-            this.safePointWaiter.wait(0, 1);
-          }
-        }
-        // Tell waiting thread we've attained safe point. Can clear this.throwable if set here
-        // because we know that next event through the ringbuffer will be going to a new WAL
-        // after we do the zigzaglatch dance.
-        this.exception = null;
-        this.zigzagLatch.safePointAttained();
-      } catch (InterruptedException e) {
-        LOG.warn("Interrupted ", e);
-        Thread.currentThread().interrupt();
-      }
-    }
-
-    /**
-     * Append to the WAL. Does all CP and WAL listener calls.
-     */
-    void append(final FSWALEntry entry) throws Exception {
-      try {
-        FSHLog.this.appendEntry(writer, entry);
-      } catch (Exception e) {
-        String msg =
-          "Append sequenceId=" + entry.getKey().getSequenceId() + ", requesting roll of WAL";
-        LOG.warn(msg, e);
-        requestLogRoll(ERROR);
-        throw new DamagedWALException(msg, e);
-      }
-    }
-
-    @Override
-    public void onStart() {
-      for (SyncRunner syncRunner : this.syncRunners) {
-        syncRunner.start();
-      }
-    }
-
-    @Override
-    public void onShutdown() {
-      for (SyncRunner syncRunner : this.syncRunners) {
-        syncRunner.interrupt();
-      }
-    }
-  }
 
   /**
    * This method gets the pipeline for the current WAL.
@@ -1211,4 +584,5 @@ public class FSHLog extends AbstractFSWAL<Writer> {
   void setWriter(Writer writer) {
     this.writer = writer;
   }
+
 }
